+++ jobapp/apps/companyapp/urls.py
@@ -10,6 +10,8 @@ urlpatterns = [
     path("companyId/", views.companyProfile, name="companyProfile"),
     path("myCompany/jobListings", views.companyJobList, name="companyJobList"),
     path("myCompany/jobListings/createJob", views.createJob, name="createJob"),
+    path("myCompany/jobListings/editJob/<int:job_id>/", views.createJob, name="editJob"),
+    path("myCompany/jobListings/deleteJob/<int:job_id>/", views.deleteJob, name="deleteJob"),
     path("myCompany/applicants", views.companyApplicants, name="companyApplicants"),
     path("myCompany/", views.companyProfileSettings, name="companyProfileSettings")
 ]
