--- jobapp/apps/jobsapp/templates/job_listings.html
@@ -1,90 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>List of Jobs | WorkIt</title>
-    <script src="https://cdn.tailwindcss.com"></script>
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.2/css/all.min.css"
-        integrity="sha512-z3gLpd7yknf1YoNbCzqRKc4qyor8gaKU1qmn+CShxbuBusANI9QpRohGBreCFkKxLhei6S9CQXFEbbKuqLg0DA=="
-        crossorigin="anonymous" referrerpolicy="no-referrer" />
-</head>
-
-<body class="bg-[#F1F8F9] min-h-screen">
-    <div class="bg-[#F1F8F9] min-h-screen">
-        {% include 'navbar.html' %}
-
-        <div class="relative rounded-xl my-5 mt-16 pt-16 p-10 w-[1100px] mx-auto">
-            <div class="text-3xl font-bold mb-6">Job Listings</div>
-            <div class="flex justify-between items-center">
-                <div class="relative inline block">
-                    <input type="text"
-                        class="border rounded-md w-64 py-2 px-4 placeholder-gray-500 focus:outline-none focus:ring focus:border-blue-300"
-                        placeholder="Search Jobs">
-                    <button class="absolute left-full top-1/2 transform -translate-y-1/2 ml-4">
-
-                  </button>
-                </div>
-                <a href="{% url "jobsapp:postJob" %}"
-                    class="bg-[#BC4749] hover:bg-green-700 text-white font-bold py-2 px-4 border border-blue-700 rounded">
-                    <i class="fa-solid fa-plus"></i>
-                    Post New Job
-                </a>
-            </div>
-
-            <div class="p-5 h-screen">
-                <div class="overflow-auto rounded-lg shadow hidden md:block">
-                    <table class="w-full">
-                      <thead class="bg-[#DFFFCB]">
-                            <tr>
-                                <th class="w-2/5 px-4 py-2 text-[#386641]">Job Title</th>
-                                <th class="w-1/5 px-4 py-2 text-[#386641]">No. of Applicants</th>
-                                <th class="w-1/10 px-4 py-2 text-[#386641]">Status</th>
-                                <th class="w-1/5 px-4 py-2 text-[#386641]">Date Posted</th>
-                                <th class="w-1/10 px-4 py-2 text-[#386641]">Actions</th>
-                            </tr>
-                        </thead>
-                        <tbody class="divide-y divide-gray-100 bg-white">
-                            {% for job in job_list %}
-                            <tr>
-                                <!--values -->
-                                <td class="border px-4 py-2">{{job.job_title}}</td>
-                                <td class="border px-4 py-2 text-center">8 Applicants</td>
-                                <td class="border px-4 py-2 text-center">
-                                  <div class="{% if job.status == 'active' %}bg-[#BCF09D] text-[#386641]{% else %}bg-[#FF8F92] text-[#BC4749]{% endif %} rounded-full">
-                                    &bull; {{ job.status|title }}
-                                </div>
-                                
-                              </td>
-                              
-                              
-                              
-                                <td class="border px-4 py-2 text-center">{{job.date_posted|date:"M j, Y"}}</td>
-                                <td class="border px-4 py-2 text-center">
-                                    <div class="justify-items-auto">
-                                        <a href="{% url 'jobsapp:editJob' job.id %}">
-                                            <i class="fa-regular fa-pen-to-square mr-2"></i>
-                                        </a>
-                                        <form action="{% url "jobsapp:deleteJob" job.id %}" method="post"
-                                            class="inline">
-                                            {% csrf_token %}
-                                            <button>
-                                                <i class="fa-regular fa-trash-can ml-2"></i>
-                                            </button>
-                                        </form>
-                                    </div>
-                                </td>
-                            </tr>
-                            {% endfor %}
-                        </tbody>
-                    </table>
-                </div>
-                <div class="grid grid-cols-1 sm:grid-cols-2 gap-4 md:hidden">
-                </div>
-            </div>
-        </div>
-    </div>
-</body>
-
-</html>
